% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duplicates.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Extract duplicate observations from data}
\usage{
get_duplicates(data, variables)
}
\arguments{
\item{data}{A data.table object.}

\item{variables}{character vector of variable names (keys) on which you want to check duplicates.}
}
\value{
A data.table object with the duplicated observations.
}
\description{
Extract duplicate observations from data
}
