% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headview.R
\name{headview}
\alias{headview}
\title{Quick function to show data}
\usage{
headview(x, numrows = 100)
}
\arguments{
\item{x}{data frame}

\item{numrows}{how many rows you want to display}
}
\value{
opens the Rstudio viewer with the first observations.
}
\description{
Quick function to show data
}
\examples{


}
