% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{idmapping}
\alias{idmapping}
\title{Use r code to connect uniprot web api}
\usage{
idmapping(query, inputid, outputid, fmt)
}
\arguments{
\item{query}{vector of protein ids}

\item{inputid}{type of input id, character}

\item{outputid}{type of output id, character}

\item{fmt}{output format}
}
\value{
a dataframe
}
\description{
You can choose one id as well as another id as output id
Know more information you can learn from uniprot documentation
}
\examples{
result <- idmapping(query=proid, inputid="ACC", outputid="P_ENTREZGENEID", fmt="fmt")

}
