% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_id.R
\name{unique_id}
\alias{unique_id}
\title{Unique ID determination in data.frame}
\usage{
unique_id(x, varnames, verbose = FALSE)
}
\description{
@description to check for unique identifiers in data.
}
\details{
@param x data frame
@param varnames character vector of variable names.
@param verbose  if TRUE the function returns a message about whether the variables are a unique identifier or not.

@import data.table
@export

@return  logical value representing whether the two variables determine a unique identifier for the data
}
