% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headview.R
\name{headview}
\alias{headview}
\title{Quick function to show data}
\usage{
headview(x, numrows = 100)
}
\description{
@description Shows in Rstudio only the top portion of a dataset.
}
\details{
@param x data frame
@param numrows how many rows you want to display

@export
}
